+++ teneto/plot/slice_plot.py
@@ -3,7 +3,7 @@ import numpy as np
 from ..utils import *
 
 
-def slice_plot(netin, ax, nodelabels='', timelabels='', timeunit='', linestyle='k-', cmap=None, nodesize=100, nodekwargs=None, edgekwargs=None):
+def slice_plot(netin, ax, nodelabels='', timelabels='', communities=None, timeunit='', linestyle='k-', cmap=None, nodesize=100, nodekwargs=None, edgekwargs=None):
     r'''
 
     Fuction draws "slice graph" and exports axis handles
@@ -19,6 +19,8 @@ def slice_plot(netin, ax, nodelabels='', timelabels='', timeunit='', linestyle='
         nodes labels. List of strings.
     timelabels : list
         labels of dimension Graph is expressed across. List of strings.
+    communities : array
+        array of size: (time) or (node,time). Nodes will be coloured accordingly.
     timeunit : string
         unit time axis is in.
     linestyle : string
@@ -137,6 +139,14 @@ def slice_plot(netin, ax, nodelabels='', timelabels='', timeunit='', linestyle='
         edgekwargs = {}
     if cmap:
         nodekwargs['cmap'] = cmap
+    if 'c' not in nodekwargs:
+        nodekwargs['c'] = posy
+    if communities is not None:
+        # check if temporal or static
+        if len(communities.shape) == 1:
+            nodekwargs['c'] = np.tile(communities, timeNum)
+        else:
+            nodekwargs['c'] = communtiies.flatten(order='F')
 
     # plt.plot(points)
     # Draw Bezier vectors around egde positions
@@ -156,7 +166,7 @@ def slice_plot(netin, ax, nodelabels='', timelabels='', timeunit='', linestyle='
     ax.get_yaxis().tick_left()
     ax.set_xlim([min(posx) - 1, max(posx) + 1])
     ax.set_ylim([min(posy) - 1, max(posy) + 1])
-    ax.scatter(posx, posy, s=nodesize, c=posy, zorder=10, **nodekwargs)
+    ax.scatter(posx, posy, s=nodesize, zorder=10, **nodekwargs)
     if timeunit != '':
         timeunit = ' (' + timeunit + ')'
     ax.set_xlabel('Time' + timeunit)