+++ teneto/temporalcommunity/__init__.py
@@ -1,0 +1,5 @@
+"""
+Functions to quantify temporal communities
+"""
+
+from .flexibility import flexibility